add initial implementation
Fixes the following issues with the previous carriersettings_extractor-based approach:
- MVNO CarrierConfig was applied on top of the base MNO CarrierConfig. Base MNO config is supposed
to be ignored in this case.
- When loading CarrierConfig, MvnoDataCase checks were performed incorrectly for GID1 MvnoDataCase
(whole string matching was used instead of prefix matching) and for IMSI MvnoDataCase ("x" and "X"
characters that are used as placeholders for a single digit weren't supported).
- The following APN fields were ignored: lingering_network_type_bitmask, always_on and mtu_v6
- Updated apns-conf.xml file wasn't automatically applied by the OS until BUILD_ID changed
(relevant only when APN extraction in carriersettings_extractor changed without stock OS build ID
change)